Refactor async exports with generic data providers
